feat(dalle image generator): included explicit enum to int cast

+++ Grains/dalle-image-generator/DalleOpenAI/DalleOpenAIImageGenerator.cs
@@ -135,7 +135,7 @@ public class DalleOpenAIImageGenerator : IImageGenerator
 
         ImageGenerationError imageGenerationError = new ImageGenerationError
         {
-            HttpStatusCode = httpStatusCode,
+            HttpStatusCode = (int)httpStatusCode,
             ImageGenerationErrorCode = imageGenerationErrorCode,
             Message = dalleOpenAiImageGenerationError!.Message
         };
